api/pyproject.toml
[project]
name = "Open-Document-Manager-API"
version = "0.1.0"
description = "Monorepo for API and AI Agent services"
requires-python = "==3.12.*"
# Common dependencies to agent and api
dependencies = [
    "loguru>=0.7.3",
]

[dependency-groups]
api = [
    "fastapi[standard]>=0.116.1",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.43",
    "sqlalchemy-utils>=0.42.0",
    "mysql-connector-python>=9.4.0",
]
dev = [
    "ruff>=0.9.0",
    "pre-commit>=4.1.0",
    "isort>=8.0.1",
]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0", # Crucial para testear endpoints async de FastAPI
    "httpx>=0.27.0",          # Necesario para el TestClient de FastAPI
]

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py312"
src = ["api", "agent"]

[tool.ruff.lint]
select = ["E", "W", "F", "FAST", "RUF", "I", "SIM", "B", "ARG", "RET", "UP", "T20", "C4"]
ignore = ["B904", "B008"]
fixable = ["ALL"]


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.pytest.ini_options]
pythonpath = ["api", "agent"]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
